Update README.md for base Rust crate and C API crate
Prior to this commit there were some errors in the CLI commands offered
to the user for following along with the READMEs which would have
prevented the manifest & header generation steps from succeeding. Also,
some clarity was missing from how to get trace data out of a system. After
this commit fixes have been applied to the CLI commands in the snippets
and additional clarifying language has been put around data collection,
visualization, and running of the tests.